+++ lib/src/widget/builder/locally_available_builder.dart
@@ -2,13 +2,13 @@
 // Use of this source code is governed by an Apache license that can be found
 // in the LICENSE file.
 
+import 'dart:developer' as dev;
 import 'dart:io';
 
 import 'package:flutter/material.dart';
 import 'package:photo_manager/photo_manager.dart';
 
 import '../../constants/extensions.dart';
-import '../../internal/methods.dart';
 import '../scale_text.dart';
 
 class LocallyAvailableBuilder extends StatefulWidget {
@@ -65,7 +65,6 @@ class _LocallyAvailableBuilderState extends State<LocallyAvailableBuilder> {
           withSubtype: true,
           progressHandler: _progressHandler,
         );
-        realDebugPrint('Produced file: $file.');
         if (file != null) {
           _isLocallyAvailable = true;
           if (mounted) {
@@ -75,7 +74,10 @@ class _LocallyAvailableBuilderState extends State<LocallyAvailableBuilder> {
       });
     }
     _progressHandler?.stream.listen((PMProgressState s) {
-      realDebugPrint('Handling progress: $s.');
+      assert(() {
+        dev.log('Handling progress: $s.');
+        return true;
+      }());
       if (s.state == PMRequestState.success) {
         _isLocallyAvailable = true;
         if (mounted) {